HW14/RestApi/pyproject.toml
[tool.poetry]
name = "restapi"
version = "0.1.0"
description = ""
authors = ["baracuda <mr.green2809@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
sqlalchemy = "^2.0.36"
pydantic = {extras = ["dotenv"], version = "^2.9.2"}
python-dotenv = "^1.0.1"
email-validator = "^2.2.0"
faker = "^30.8.2"
flask-jwt-extended = "^4.6.0"
werkzeug = "^3.1.3"
jose = "^1.0.0"
passlib = "^1.7.4"
pydantic-settings = "^2.6.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.17"
alembic = "^1.14.0"
pyjwt = "^2.10.0"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
redis = "^5.2.0"
cloudinary = "^1.41.0"
psycopg2-binary = "^2.9.10"
fastapi-cache = "^0.1.0"
fastapi-cache2 = {extras = ["dynamodb"], version = "^0.2.2"}
authlib = "^1.3.2"
python-magic = "^0.4.27"
psycopg2 = "^2.9.5"
sphinx = "^8.1.3"
pytest = "^8.3.3"
httpx = "^0.28.0"
docker = "^7.1.0"
setuptools = "^75.6.0"
pytest-asyncio = "^0.24.0"
aiosqlite = "^0.20.0"
pytest-cov = "^6.0.0"




[tool.poetry.group.dev.dependencies]
uvicorn = "^0.32.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

